% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_firesting.R
\encoding{UTF-8}
\name{import_firesting}
\alias{import_firesting}
\title{Import data from Pyro Oxygen Logger}
\usage{
import_firesting(
  file,
  o2_unit = "percent_a.s.",
  date = "\%m/\%d/\%Y \%X",
  overwrite_sal = NULL,
  keep_metadata = FALSE,
  drop_channels = TRUE,
  split_channels = FALSE
)
}
\arguments{
\item{file}{a character string. The filepath for the file to be read.}

\item{o2_unit}{a character string. The unit of O2 measurement to be output in the data frame. Options are described in \code{\link{conv_o2}}.}

\item{date}{a character string. The date format to be passed to \code{\link{strptime}}.}

\item{overwrite_sal}{Default \code{NULL}. To overwrite the salinity value(s) from calibration, enter a single numeric value for all channels or a numeric vector with values for each channel. Salinity of water sample (psu).}

\item{keep_metadata}{logical. Should metadata from the file be returned as extra columns in the returned data frame? Default is \code{FALSE}.}

\item{drop_channels}{logical. Should channels without any O2 data be dropped? Default is \code{TRUE}.}

\item{split_channels}{logical. Should a list of data frames be returned with a separate data frame for each channel? Default is \code{FALSE}.}
}
\value{
A data frame (or list of data frames) is returned.
\describe{
\item{TIME}{Date and time, POSIXlt format.}
\item{DURATION}{Duration of measurement trial (minutes).}
\item{CH_X_O2}{Oxygen measurement in desired unit as determined by \code{o2_unit}.}
\item{CH_X_TEMP}{Temperature recorded or defined at beginning of measurement trial.}
\item{CH_X_SAL}{Salinity (psu).}
\item{...}{Channel columns (CH_...) are repeated for each channel.}
\item{COMMENT}{Comments from FireSting file.}
}
If \code{keep_metadata = TRUE}, then the following columns are appended to the returned data frame:
\describe{
\item{ATM_PRES}{Atmospheric pressure (mbar).}
\item{HUMIDITY}{Relative humidity (\% RH).}
\item{PROBE_TEMP}{Probe temperature.}
\item{INTERNAL_TEMP}{Transmitter internal temperature.}
\item{ANALOG_IN}{Voltage input from the extension port (mV).}
\item{CH_X_PHASE}{Phase recorded. Phase is inversely related to O2.}
\item{CH_X_INTENSITY}{Intensity is an indicator of the quality of the signal. A low intensity warning is produced by the transmitter below 10 mV.}
\item{CH_X_AMB_LIGHT}{Ambient light on the sensor. Expressed in mV.}
}

If \code{split_channels = TRUE}, then "\code{CH_X_}" is removed from the column names and multiple data frames are returned in a named list.
}
\description{
Imports the standard txt file output from Pyroscience's deprecated Pyro Oxygen Logger software and converts the data into one or more data frames. If using the newer Pyroscience Workbench software, use \code{\link{import_pyroscience_workbench}} instead.
}
\details{
The following FireSting fiber optic O2 transmitters are supported:
\itemize{
\item{FireStingO2}{}
\item{FireStingO2 (1st generation)}{}
}
If you would like support for the Piccolo2, FireStingO2-Mini, TeX4, or any OEM instruments, email me a data file from the device.
}
\note{
Oxygen conversions are estimates based on the \code{\link[marelac]{marelac}} package.
}
\examples{
\dontrun{
file <- system.file('extdata', 'pyro_oxygen_logger_file.txt', package = 'respirometry')
import_firesting(file, o2_unit = 'umol_per_l')

# I want each channel as a separate data frame.
data_list <- import_firesting(file, split_channels = TRUE)
data_list$CH_3 # here's the channel 3 data frame.
}

}
\seealso{
\code{\link{import_pyroscience_workbench}}, \code{\link{import_presens}}, \code{\link{import_witrox}}, \code{\link{conv_o2}}
}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}
