% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pH.R
\encoding{UTF-8}
\name{predict_pH}
\alias{predict_pH}
\title{Predict pH post-respiration}
\usage{
predict_pH(start_o2 = 100, end_o2, start_pH, temp = 25, sal = 35,
  RQ = 1, TA = NULL, all_carb = FALSE)
}
\arguments{
\item{start_o2}{pO2 at the start of the measurement (\% air saturation). Default is 100\% air saturation.}

\item{end_o2}{pO2 at the end of the measurment (\% air saturation).}

\item{start_pH}{seawater pH (total scale) at the start of the measurement.}

\item{temp}{temperature (°C). Default is 25 °C.}

\item{sal}{salinity (psu). Default is 35 psu. If \code{sal} < 26 psu, then \code{TA} must be provided.}

\item{RQ}{respiratory quotient: ratio of CO2 produced / O2 consumed. Default is 1.}

\item{TA}{(optional) total alkalinity (umol / kg). If undefined TA is estimated from salinity using \code{\link{guess_TA}}.}

\item{all_carb}{logical. Should all carbonate chemistry parameters be returned? Default is FALSE.}
}
\value{
If \code{all_carb} is \code{FALSE}, then a list of the predicted pH (total scale) at the end of the measurement and the predicted pCO2 (uatm) are returned. If \code{all_carb} is \code{TRUE}, then the predicted carbonate chemistry parameters are returned from \code{\link[seacarb]{carb}}.
}
\description{
Predicts the pH of seawater after a defined amount of oxygen consumption.
}
\details{
Given a known amount of oxygen consumed and an estimated respiratory quotient (see \code{\link{Q10}}), the amount of CO2 produced can be estimated. From this CO2 production estimate, the carbonate chemistry of the seawater can be estimated. Atmospheric pressure is assumed.
}
\examples{
predict_pH(end_o2 = 75, start_pH = 8.1)
predict_pH(start_o2 = 75, end_o2 = 50, start_pH = 7.96, temp = 15, sal = 33, RQ = 0.88)

# I know pH at the end was 7.8, but what was pH at the beginning?
predict_pH(start_o2 = 75, end_o2 = 100, start_pH = 8.013536) # reverse the order

}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}
\seealso{
\code{\link[seacarb]{carb}}, \code{\link{guess_TA}}
}

