% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respirometry.R
\docType{package}
\encoding{UTF-8}
\name{respirometry}
\alias{respirometry}
\title{Tools for Conducting Respirometry Experiments}
\description{
Provides tools to enable the researcher to more precisely conduct respirometry experiments. Strong emphasis is on aquatic respirometry. Tools focus on helping the researcher setup and conduct experiments. Analysis of the resulting data is not a focus since analyses are often specific to a particular setup, and thus are better created by the researcher individually. This package provides tools for intermittent, flow-through, and closed respirometry techniques.
}
\author{
Matthew A. Birk, \email{matthewabirk@gmail.com}
}
